Add loop marker nodes to fix analysis error
The Nullable dereference analyzer revealed a problem where the CFG for
loop statements was not sufficiently precise. It was not possible to
tell the difference between branches leaving the loop condition, and
branches entering the loop if there were no CFG nodes generated inside
the loop.

Updated tests to account for the new CFG marker nodes. Added regression
tests for analysis failures caused by the previous less-precise CFGs.