Assume all method calls are null guard equivalent
The NullableDereference analyzer now considers a method call with an argument x
as being a null guard on argument x.  This removes some false positives caused
by calling precondition-like methods that throw an exception if an argument is
null. Such methods could be used as effective null guards, aborting the current
control flow if some parameter is null.

This change reduces the precision of the NullableDereference analyzer, but
hopefully the reduced false positive rate outweighs that drawback.