Update to ExtendJ 8.0.1-126-ga069f0d
This updates the third-party ExtendJ dependency to version
8.0.1-126-ga069f0d. ExtendJ was refactored so that VariableDeclaration
was replaced by VariableDeclarator, and FieldDeclaration was replaced by
FieldDeclarator. Updating SimpleCfg to work with this change is easy -
simply switch to using the replacement AST nodes.

Line number information for MethodAccess nodes changed slightly - it now
points to the beginning of the qualifier expression, rather than the
method name. Two test cases were affected by this and updated to expect
the new source location in the test output.

ExtendJ now has a simpler configuration for default scanner/parsers,
however this introduced a dependency on having a class named exactly
org.extendj.parser.JavaParser. We thus need to generate this class
during the build.

fixes #20